release: Allow specimenCOllectedVOlume in mls to be float value

+++ backend/src/main/java/com/Adapter/icare/Dtos/LabRequestDetailsDTO.java
@@ -24,7 +24,7 @@ public class LabRequestDetailsDTO {
     private CodeDTO typeOfSpecimen;
     private CodeDTO specimenSite;
     private SpecimentCollectedFrom specimenCollectedFrom;
-    private Integer specimenCollectionVolumeInMl;
+    private Float specimenCollectionVolumeInMl;
     private String specimenCollectorName;
     private String specimenCollectorContactNumber;
     private Instant dateTimeSpecimenCollected;
